All the xforms now have maker functions that take care of initialization.
This is to make the maker functions more efficient. Instead of re-reading
and processing the "config" for the xforms every single time, the maker
functions do it once and return an "initialized" transform function.

Also, all the xforms now just delegate to the maker functions, in the
spirit of DRY. So the unit tests all pass now :P

Also, deleted the registries from reader/json/core.clj and xform/core.clj,
since we aren't using those registries anymore.